iot_kafka_producer/requirements.txt
paho-mqtt
kafka-python
pydantic[dotenv]
pydantic==1.*